requirements.txt
python-telegram-bot==20.7
pydub
openai
whisper
httpx
